feat: use floats for position and width / height instead of ints

+++ src/objects/object.cpp
@@ -9,14 +9,14 @@ Object::~Object() {
 		SDL_DestroySurface(this->_texture);
 }
 
-void Object::set_position(int x, int y, int z)
+void Object::set_position(float x, float y, float z)
 {
 	this->_x = x;
 	this->_y = y;
 	this->_z = z;
 }
 
-void Object::set_size(int width, int height)
+void Object::set_size(float width, float height)
 {
 	this->_width = width;
 	this->_height = height;
@@ -41,10 +41,10 @@ void Object::draw(Window *window)
 {
 	SDL_Texture	*tex;
 	SDL_FRect	rect;
-	rect.x = (float) this->_x;
-	rect.y = (float) this->_y;
-	rect.w = (float) this->_width;
-	rect.h = (float) this->_height;
+	rect.x = this->_x;
+	rect.y = this->_y;
+	rect.w = this->_width;
+	rect.h = this->_height;
 	tex = SDL_CreateTextureFromSurface(window->get_renderer(), this->_texture);
 	if (this->_rotation != 0)
 		SDL_RenderTextureRotated(window->get_renderer(), tex, nullptr, &rect, this->_rotation, nullptr, SDL_FLIP_NONE);
